feat: added hash64 based on murmur3
