+++ src/impressions/hasher/__tests__/hashImpression32.spec.js
@@ -1,7 +1,7 @@
 import tape from 'tape-catch';
-import { hashImpression128 } from '../hashImpression128';
+import { hashImpression32 } from '../hashImpression32';
 
-tape('Hasher 128 / Impression Hasher Works', assert => {
+tape('Hasher 32 / Impression Hasher Works', assert => {
   const imp1 = {
     feature: 'someFeature',
     keyName: 'someKey',
@@ -18,36 +18,36 @@ tape('Hasher 128 / Impression Hasher Works', assert => {
     label: 'someLabel',
     treatment: 'someTreatment',
   };
-  assert.equal(hashImpression128(imp1), hashImpression128(imp2));
+  assert.equal(hashImpression32(imp1), hashImpression32(imp2));
 
   // Different feature
   imp2.feature = 'someOtherFeature';
-  assert.notEqual(hashImpression128(imp1), hashImpression128(imp2));
+  assert.notEqual(hashImpression32(imp1), hashImpression32(imp2));
 
   // Different key
   imp2.feature = imp1.feature;
   imp2.keyName = 'someOtherKey';
-  assert.notEqual(hashImpression128(imp1), hashImpression128(imp2));
+  assert.notEqual(hashImpression32(imp1), hashImpression32(imp2));
 
   // Different changeNumber
   imp2.keyName = imp1.keyName;
   imp2.changeNumber = 456;
-  assert.notEqual(hashImpression128(imp1), hashImpression128(imp2));
+  assert.notEqual(hashImpression32(imp1), hashImpression32(imp2));
 
   // Different label
   imp2.changeNumber = imp1.changeNumber;
   imp2.label = 'someOtherLabel';
-  assert.notEqual(hashImpression128(imp1), hashImpression128(imp2));
+  assert.notEqual(hashImpression32(imp1), hashImpression32(imp2));
 
   // Different Treatment
   imp2.label = imp1.label;
   imp2.treatment = 'someOtherTreatment';
-  assert.notEqual(hashImpression128(imp1), hashImpression128(imp2));
+  assert.notEqual(hashImpression32(imp1), hashImpression32(imp2));
 
   assert.end();
 });
 
-tape('Hasher 128 / Impression Hasher Does Not Crash', assert => {
+tape('Hasher 32 / Impression Hasher Does Not Crash', assert => {
   const imp1 = {
     feature: 'someFeature',
     keyName: 'someKey',
@@ -57,18 +57,16 @@ tape('Hasher 128 / Impression Hasher Does Not Crash', assert => {
   };
 
   imp1.keyName = null;
-  assert.isNot(hashImpression128(imp1), null);
+  assert.isNot(hashImpression32(imp1), null);
 
   imp1.changeNumber = null;
-  assert.isNot(hashImpression128(imp1), null);
+  assert.isNot(hashImpression32(imp1), null);
 
   imp1.label = null;
-  assert.isNot(hashImpression128(imp1), null);
+  assert.isNot(hashImpression32(imp1), null);
 
   imp1.treatment = null;
-  assert.isNot(hashImpression128(imp1), null);
-
-  assert.is(hashImpression128(null), null);
+  assert.isNot(hashImpression32(imp1), null);
 
   assert.end();
 });
