+++ src/impressions/hasher/buildKey.js
@@ -1,0 +1,3 @@
+export function buildKey(impression) {
+  return `${impression.keyName}:${impression.feature}:${impression.treatment}:${impression.label}:${impression.changeNumber}`;
+}
