+++ src/impressions/hasher/hashImpression32.js
@@ -1,6 +1,6 @@
 import murmur from '../../engine/engine/murmur3/murmur3';
-import { buildKey } from './';
+import { buildKey } from './buildKey';
 
 export function hashImpression32(impression) {
-  return impression ? murmur.hash(buildKey(impression)).toString() : null;
+  return murmur.hash(buildKey(impression));
 }
